Remove unnecessary cast to `float`
This fixes Sonar finding:

> Redundant casts should not be used [java:S1905](https://rules.sonarsource.com/java/RSPEC-1905/)

+++ mq/main/mq-broker/mqbroker-core/src/main/java/com/sun/messaging/jmq/jmsserver/core/Destination.java
@@ -2310,7 +2310,7 @@ public abstract class Destination implements DestinationSpi, Serializable, com.s
         if (maxc <= 0) {
             return 0;
         }
-        return ((float) destMessages.size() / (float) maxc) * 100;
+        return ((float) destMessages.size() / maxc) * 100;
     }
 
     public abstract int getUnackSize();
